+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Added: [#5557](https://github.com/ethereum/aleth/pull/5557) Improved debug logging of full sync.
 - Changed: [#5559](https://github.com/ethereum/aleth/pull/5559) Update peer validation error messages.
 - Fixed: [#5562](https://github.com/ethereum/aleth/pull/5562) Don't send header request messages to peers that haven't sent us Status yet.
+- Added: [#5564](https://github.com/ethereum/aleth/pull/5564) Improved help menu of Aleth by adding list of channels.
 
 ## [1.6.0] - Unreleased